pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "privacy-service"
version = "0.1.0"
description = "PII detection and anonymization library with CLI using Microsoft Presidio and ai4privacy"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Privacy Service Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Text Processing",
]
keywords = ["pii", "privacy", "anonymization", "presidio", "nlp", "data-protection"]

dependencies = [
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "typer[all]>=0.9.0",
    "pandas>=2.0.0",
    "pypdf>=4.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "chardet>=5.0.0",
    "sentencepiece>=0.2.1",
    "protobuf>=6.33.5",
    "ai4privacy>=0.3.3",
    "spacy-transformers>=1.3.9",
]


[project.scripts]
privacy-service = "privacy_service.cli.main:app"
lint = "privacy_service.tools:lint"
format = "privacy_service.tools:format"
type-check = "privacy_service.tools:type_check"

[project.urls]
Homepage = "https://github.com/yourusername/privacy-service"
Documentation = "https://github.com/yourusername/privacy-service"
Repository = "https://github.com/yourusername/privacy-service"
Issues = "https://github.com/yourusername/privacy-service/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/privacy_service"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=privacy_service --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[dependency-groups]
dev = [
    "datasets>=4.5.0",
    "ipykernel>=6.31.0",
    "ipython>=8.18.1",
    "mypy>=1.19.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=4.1.0",
    "ruff>=0.15.0",
    "types-pyyaml>=6.0.12.20250915",
]
